requirements.txt
absl-py==2.1.0
addict==2.4.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.8.0
attrs==24.3.0
basicsr==1.4.2
blinker==1.9.0
cachetools==5.5.1
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
cmake==3.31.4
contourpy==1.3.1
cycler==0.12.1
exceptiongroup==1.2.2
facexlib==0.2.5
fastapi==0.115.7
filelock==3.17.0
filterpy==1.4.5
fonttools==4.55.5
fsspec==2024.12.0
future==1.0.0
gfpgan==1.3.5
gitdb==4.0.12
GitPython==3.1.44
grpcio==1.70.0
h11==0.14.0
idna==3.10
imageio==2.37.0
Jinja2==3.1.5
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.8
lazy_loader==0.4
lit==18.1.8
llvmlite==0.44.0
lmdb==1.6.2
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
mdurl==0.1.2
mpmath==1.3.0
narwhals==1.23.0
networkx==3.4.2
numba==0.61.0
numpy==1.24.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu11==10.2.10.91
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu11==11.7.4.91
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu11==2.14.3
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu11==11.7.91
nvidia-nvtx-cu12==12.4.127
opencv-python==4.11.0.86
packaging==24.2
pandas==2.2.3
pillow==11.1.0
platformdirs==4.3.6
protobuf==5.29.3
pyarrow==19.0.0
pydantic==2.10.6
pydantic_core==2.27.2
pydeck==0.9.1
Pygments==2.19.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
realesrgan==0.3.0
referencing==0.36.1
requests==2.32.3
rich==13.9.4
rpds-py==0.22.3
scikit-image==0.25.0
scipy==1.15.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
starlette==0.45.3
streamlit==1.41.1
sympy==1.13.1
tb-nightly==2.19.0a20250123
tenacity==9.0.0
tensorboard-data-server==0.7.2
tifffile==2025.1.10
toml==0.10.2
tomli==2.2.1
torch==2.0.0
torchvision==0.15.1
tornado==6.4.2
tqdm==4.67.1
triton==2.0.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0
watchdog==6.0.0
Werkzeug==3.1.3
yapf==0.43.0
